backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
<<<<<<< HEAD
python-dotenv==1.0.1
=======
python-dotenv==1.0.1
google-auth==2.33.0
requests==2.32.3
>>>>>>> diego/main
